Allow loading multiple series with DICOMweb
`WsiDicom.open_web()` was modified to be able to either take a single
series uid (as before) or a list of series uids. If a list of series
uids is passed, their instances are all loaded together. This can be
useful for cases where, for instance, different optical paths are located
in different series.

This also loosens the UID matching to only check the frame of references.

This also adds `TotalPixelMatrixOriginSequence` matching when comparing
datasets.

Fixes: #118

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>